requirements.txt
Flask==3.0.2
requests==2.31.0
beautifulsoup4==4.12.3
gunicorn==21.2.0
gnews
newspaper3k


